+++ BusyBeaverFinder/Utils.cpp
@@ -33,8 +33,8 @@ int makePowerOfTwo(int val) {
 
 InstructionPointer nextInstructionPointer(ProgramPointer pp) {
     return InstructionPointer {
-        .col = pp.p.col + dx[(int)pp.dir],
-        .row = pp.p.row + dy[(int)pp.dir]
+        .col = (int8_t)(pp.p.col + dx[(int)pp.dir]),
+        .row = (int8_t)(pp.p.row + dy[(int)pp.dir])
     };
 }
